refactor: improve module organization and separation of concerns
Split grab-bag modules into focused, single-purpose modules:

- Create branch_deletion.rs: domain logic for integration checking and branch
  deletion (moved from handlers.rs)
- Create hook_commands.rs: all `wt hook` subcommand handlers (moved from
  standalone.rs)
- Rename standalone.rs → step_commands.rs: now focused solely on merge workflow
  steps (commit, squash, rebase)
- Move build_remove_command to process.rs: shell command construction belongs
  with other process management code

This improves cohesion by keeping output/handlers.rs focused on presentation
and moving domain logic to appropriate command modules.
